now able to sign packages with old pk
