+++ src/Core/Crypto.cs
@@ -70,12 +70,12 @@ public static class Crypto
     }
     
     [SupportedOSPlatform("windows")]
-    public static byte[] SignData(string? password, byte[] payloadZip)
+    public static byte[] SignData(string? password, byte[] payloadZip, string? customPrivateKeyPath = null)
     {
         byte[]? pass = null;
         if(password != null)
             pass = Encoding.UTF8.GetBytes(password);
-        var privateKeyBytes = File.ReadAllBytes(GetKeyPath(CurrentPrivateKeyFilename));
+        var privateKeyBytes = File.ReadAllBytes(GetKeyPath(customPrivateKeyPath ?? CurrentPrivateKeyFilename));
         var rawPrivate = ProtectedData.Unprotect(privateKeyBytes, pass
             , DataProtectionScope.CurrentUser);
